+++ actionpack/lib/action_controller/railtie.rb
@@ -15,6 +15,7 @@ module ActionController
     config.action_controller.raise_on_open_redirects = false
     config.action_controller.log_query_tags_around_actions = true
     config.action_controller.wrap_parameters_by_default = false
+    config.action_controller.html_assertions = :rails_dom_testing
 
     config.eager_load_namespaces << AbstractController
     config.eager_load_namespaces << ActionController
@@ -85,7 +86,8 @@ module ActionController
           :action_on_unpermitted_parameters,
           :always_permitted_parameters,
           :wrap_parameters_by_default,
-          :allow_deprecated_parameters_hash_equality
+          :allow_deprecated_parameters_hash_equality,
+          :html_assertions
         )
 
         filtered_options.each do |k, v|
@@ -141,8 +143,21 @@ module ActionController
     end
 
     initializer "action_controller.test_case" do |app|
+      html_assertions = app.config.action_controller.delete(:html_assertions)
+
       ActiveSupport.on_load(:action_controller_test_case) do
         ActionController::TestCase.executor_around_each_request = app.config.active_support.executor_around_test_case
+
+        case html_assertions
+        when :capybara
+          include ActionView::CapybaraAssertions
+        when :rails_dom_testing
+          include ActionView::RailsDomTestingAssertions
+        when :none
+          # do nothing
+        else
+          raise ArgumentError.new("unrecognized value #{assertions.inspect} for config.action_controller.html_assertions")
+        end
       end
     end
   end